Game Of NIM
There are two modes in this game. In the first mode, the computer plays against a human
player. The computer can be Novice, and in its turn just takes a random number of marbles, or
computer can be Smart, and in its turn uses a smart strategy to take the best number of marbles trying
to win the game at the end.
Note that in any case, computer will follow the game rules. In the second mode two human players can
play against each other.